Move `systemProp` function to the common module; only the JVM platform implements it properly while others always use provided default value.
