+++ kotlinx-coroutines-core/native/src/internal/SystemProps.kt
@@ -1,0 +1,7 @@
+/*
+ * Copyright 2016-2019 JetBrains s.r.o. Use of this source code is governed by the Apache 2.0 license.
+ */
+
+package kotlinx.coroutines.internal
+
+internal actual fun systemProp(propertyName: String): String? = null
